pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "gopnik"
dynamic = ["version"]
description = "AI-powered forensic-grade deidentification toolkit"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Gopnik Development Team", email = "dev@gopnik.ai"}
]
maintainers = [
    {name = "Gopnik Development Team", email = "dev@gopnik.ai"}
]
keywords = ["pii", "deidentification", "redaction", "privacy", "ai", "forensic"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Legal Industry",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing",
    "Topic :: Office/Business",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "click>=8.0",
    "cryptography>=3.4.8",
    "pillow>=9.0.0",
    "colorama>=0.4.4",
    "tqdm>=4.64.0",
]

[project.optional-dependencies]
web = [
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.0",
]
ai = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.30.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "build>=0.10.0",
    "twine>=4.0.0",
]
all = [
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.30.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
]

[project.urls]
Homepage = "https://github.com/happy2234/gopnik"
Documentation = "https://happy2234.github.io/gopnik/"
Repository = "https://github.com/happy2234/gopnik"
"Bug Reports" = "https://github.com/happy2234/gopnik/issues"
Changelog = "https://github.com/happy2234/gopnik/releases"
Wiki = "https://github.com/happy2234/gopnik/wiki"

[project.scripts]
gopnik = "gopnik.interfaces.cli.main:main"
gopnik-api = "gopnik.interfaces.api.cli:serve"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
gopnik = ["py.typed"]

[tool.setuptools_scm]
write_to = "src/gopnik/_version.py"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false